Move classes to separate files, fix naming
