+++ API/src/Modules/XpSystems/Momentum.XpSystems.Api/ViewModels/XpSystemViewModel.cs
@@ -1,14 +1,16 @@
 ﻿using Momentum.XpSystems.Core.Models;
 using System.Text.Json.Serialization;
+using Momentum.XpSystems.Core.Models.Cosmetic;
+using Momentum.XpSystems.Core.Models.Rank;
 
 namespace Momentum.XpSystems.Api.ViewModels
 {
     public class XpSystemViewModel
     {
         [JsonPropertyName("rankXP")]
-        public RankXP RankXP { get; set; }
+        public RankXp RankXP { get; set; }
 
         [JsonPropertyName("cosXP")]
-        public CosmeticXP CosXP { get; set; }
+        public CosmeticXp CosXP { get; set; }
     }
 }
