+++ API/src/Modules/XpSystems/Momentum.XpSystems.Core/Models/Cosmetic/Completions.cs
@@ -1,0 +1,8 @@
+﻿namespace Momentum.XpSystems.Core.Models.Cosmetic
+{
+    public class Completions
+    {
+        public Unique Unique { get; set; }
+        public Repeat Repeat { get; set; }
+    }
+}
