+++ API/src/Modules/XpSystems/Momentum.XpSystems.Core/Models/Cosmetic/CosmeticXp.cs
@@ -1,0 +1,8 @@
+﻿namespace Momentum.XpSystems.Core.Models.Cosmetic
+{
+    public class CosmeticXp
+    {
+        public Levels Levels { get; set; }
+        public Completions Completions { get; set; }
+    }
+}
