+++ API/src/Modules/XpSystems/Momentum.XpSystems.Core/Models/Cosmetic/Repeat.cs
@@ -1,0 +1,7 @@
+﻿namespace Momentum.XpSystems.Core.Models.Cosmetic
+{
+    public class Repeat
+    {
+        public RepeatTierScale TierScale { get; set; }
+    }
+}
