+++ API/src/Modules/XpSystems/Momentum.XpSystems.Core/Models/Cosmetic/RepeatTierScale.cs
@@ -1,0 +1,10 @@
+﻿namespace Momentum.XpSystems.Core.Models.Cosmetic
+{
+    public class RepeatTierScale
+    {
+        public int Linear { get; set; }
+        public int Staged { get; set; }
+        public int Stages { get; set; }
+        public int Bonus { get; set; }
+    }
+}
