+++ API/src/Modules/XpSystems/Momentum.XpSystems.Core/Models/Cosmetic/Unique.cs
@@ -1,0 +1,7 @@
+﻿namespace Momentum.XpSystems.Core.Models.Cosmetic
+{
+    public class Unique
+    {
+        public UniqueTierScale TierScale { get; set; }
+    }
+}
