+++ API/src/Modules/XpSystems/Momentum.XpSystems.Core/Models/Cosmetic/UniqueTierScale.cs
@@ -1,0 +1,8 @@
+﻿namespace Momentum.XpSystems.Core.Models.Cosmetic
+{
+    public class UniqueTierScale
+    {
+        public int Linear { get; set; }
+        public int Staged { get; set; }
+    }
+}
