+++ API/src/Modules/XpSystems/Momentum.XpSystems.Core/Models/Rank/Formula.cs
@@ -1,0 +1,8 @@
+﻿namespace Momentum.XpSystems.Core.Models.Rank
+{
+    public class Formula
+    {
+        public int A { get; set; }
+        public int B { get; set; }
+    }
+}
