+++ API/src/Modules/XpSystems/Momentum.XpSystems.Core/Models/Rank/RankXp.cs
@@ -1,0 +1,9 @@
+﻿namespace Momentum.XpSystems.Core.Models.Rank
+{
+    public class RankXp
+    {
+        public Top10 Top10 { get; set; }
+        public Formula Formula { get; set; }
+        public Groups Groups { get; set; }
+    }
+}
