+++ API/src/Modules/XpSystems/Momentum.XpSystems.Core/Models/Rank/Top10.cs
@@ -1,0 +1,8 @@
+﻿namespace Momentum.XpSystems.Core.Models.Rank
+{
+    public class Top10
+    {
+        public int WorldRecordPoints { get; set; }
+        public float[] RankPercentages { get; set; }
+    }
+}
